+++ src/azure-cli/azure/cli/command_modules/storage/tests/latest/test_storage_blob_live_scenarios.py
@@ -93,78 +93,3 @@ class StorageBlobUploadLiveTests(LiveScenarioTest):
             self.assertTrue(os.path.isfile(downloaded), 'The file is not downloaded.')
             self.assertEqual(file_size_kb * 1024, os.stat(downloaded).st_size,
                              'The download file size is not right.')
-
-    @ResourceGroupPreparer(name_prefix="storage_blob_restore", location="centraluseuap")
-    @StorageAccountPreparer(name_prefix="restore", kind="StorageV2", sku='Standard_LRS', location="centraluseuap")
-    def test_storage_blob_restore(self, resource_group, storage_account):
-        import time
-        self.cmd('storage account blob-service-properties update --enable-change-feed --enable-delete-retention --delete-retention-days 2 --enable-versioning -n {sa}')\
-            .assert_with_checks(JMESPathCheck('changeFeed.enabled', True),
-                                JMESPathCheck('deleteRetentionPolicy.enabled', True),
-                                JMESPathCheck('deleteRetentionPolicy.days', 2))
-
-        c1 = self.create_random_name(prefix='containera', length=24)
-        c2 = self.create_random_name(prefix='containerb', length=24)
-        b1 = self.create_random_name(prefix='blob1', length=24)
-        b2 = self.create_random_name(prefix='blob2', length=24)
-        b3 = self.create_random_name(prefix='blob3', length=24)
-        b4 = self.create_random_name(prefix='blob4', length=24)
-
-        local_file = self.create_temp_file(256)
-
-        account_key = self.cmd('storage account keys list -n {} -g {} --query "[0].value" -otsv'
-                               .format(storage_account, resource_group)).output
-
-        # Prepare containers and blobs
-        for container in [c1, c2]:
-            self.cmd('storage container create -n {} --account-name {} --account-key {}'.format(
-                container, storage_account, account_key)) \
-                .assert_with_checks(JMESPathCheck('created', True))
-            for blob in [b1, b2, b3, b4]:
-                self.cmd('storage blob upload -c {} -f "{}" -n {} --account-name {} --account-key {}'.format(
-                    container, local_file, blob, storage_account, account_key))
-            self.cmd('storage blob list -c {} --account-name {} --account-key {}'.format(
-                container, storage_account, account_key)) \
-                .assert_with_checks(JMESPathCheck('length(@)', 4))
-
-            self.cmd('storage container delete -n {} --account-name {} --account-key {}'.format(
-                container, storage_account, account_key)) \
-                .assert_with_checks(JMESPathCheck('deleted', True))
-
-        time.sleep(60)
-        # Enable Restore Policy
-        self.cmd('storage account blob-service-properties update --enable-restore-policy --restore-days 1 -n {sa}')
-        self.cmd('storage account blob-service-properties show -n {sa}') \
-            .assert_with_checks(JMESPathCheck('restorePolicy.enabled', True),
-                                JMESPathCheck('restorePolicy.days', 1),
-                                JMESPathCheckExists('minRestoreTime'))
-        # Restore blobs, with specific ranges
-        time_to_restore = (datetime.utcnow() + timedelta(seconds=-5)).strftime('%Y-%m-%dT%H:%MZ')
-
-        # c1/b1 -> c1/b2
-        start_range = '/'.join([c1, b1])
-        end_range = '/'.join([c1, b2])
-        self.cmd('storage blob restore -t {} -r {} {} --account-name {} -g {}'.format(
-            time_to_restore, start_range, end_range, storage_account, resource_group), checks=[
-            JMESPathCheck('status', 'Complete'),
-            JMESPathCheck('parameters.blobRanges[0].startRange', start_range),
-            JMESPathCheck('parameters.blobRanges[0].endRange', end_range)])
-
-        self.cmd('storage blob restore -t {} -r {} {} --account-name {} -g {} --no-wait'.format(
-            time_to_restore, start_range, end_range, storage_account, resource_group))
-
-        time.sleep(90)
-
-        time_to_restore = (datetime.utcnow() + timedelta(seconds=-5)).strftime('%Y-%m-%dT%H:%MZ')
-        # c1/b2 -> c2/b3
-        start_range = '/'.join([c1, b2])
-        end_range = '/'.join([c2, b3])
-        self.cmd('storage blob restore -t {} -r {} {} --account-name {} -g {}'.format(
-            time_to_restore, start_range, end_range, storage_account, resource_group), checks=[
-            JMESPathCheck('status', 'Complete'),
-            JMESPathCheck('parameters.blobRanges[0].startRange', start_range),
-            JMESPathCheck('parameters.blobRanges[0].endRange', end_range)])
-
-        time.sleep(120)
-        self.cmd('storage blob restore -t {} --account-name {} -g {} --no-wait'.format(
-            time_to_restore, storage_account, resource_group))
